requirements.txt
matplotlib
numpy
opencv_python
Pillow
PyYAML
rasterio
torch
torchvision
tqdm
